Don't set a default minimum healthy percentage of 100% anymore, so either the default of 90% will be used, or the percentage set in the instance maintenance policy for the ASG.
Introduce the properties `min_healthy_percentage` and `max_healthy_percentage`
on the autoscale definitions for tuning these settings.

The property `healthy_percentage` is still supported but
will give a deprecation warning, and will be removed somewhere in the future.

Gem `aws-sdk-autoscaling` >= 1.100.0 is now required to support
setting the maximum healthy percentage.